% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_functions.R
\name{blank_plot}
\alias{blank_plot}
\title{Make a blank plot with ggplot2}
\usage{
blank_plot()
}
\value{
A \code{ggplot} object.
}
\description{
Generates a blank plot, useful for making white space in combined plots using e.g. cowplot.
}
