% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendro.R
\name{dendro}
\alias{dendro}
\title{Make a dendrogram from an hclust object}
\usage{
dendro(
  clust,
  edge = c("bottom", "left", "top", "right"),
  plot_margin = sapply(c("top", "right", "bottom", "left"), function(x) ifelse(x ==
    edge, 0, 5.5)),
  size = 0.5
)
}
\arguments{
\item{clust}{An \code{hclust} object.}

\item{edge}{The edge of the plot window to which to align the dendrogram, possible values being \code{'bottom'}, \code{'left'}, \code{'top'} and \code{'right'}. For example, if the dendrogram is to be plotted above a heatmap, use \code{edge = 'bottom'}. Default: \code{'bottom'}.}

\item{plot_margin}{A numeric vector of margin sizes in pts, in the order top, right, bottom, left. Default: 0 for \code{edge}, 5.5 for all other margins.}

\item{size}{The weight of the lines.}
}
\value{
A ggplot object.
}
\description{
Makes a dendrogram aligned to a chosen edge of the plot area. This function is useful for making a dendrogram to align next to a heatmap. It is used in the \code{heat_map} function.
}
