% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_functions.R
\name{ggplot_colours}
\alias{ggplot_colours}
\title{ggplot2 default colours}
\usage{
ggplot_colours(n)
}
\arguments{
\item{n}{The number of colours to be returned.}
}
\value{
A character vector of hexadecimal colour codes.
}
\description{
Generates colours like the default colours used in ggplot2, by taking points evenly distributed around a colour wheel.
}
