% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_functions.R
\name{headl}
\alias{headl}
\alias{headr}
\alias{taill}
\alias{tailr}
\title{Head/tail left/right}
\usage{
headl(x, m = 6L, n = 6L)

headr(x, m = 6L, n = 6L)

taill(x, m = 6L, n = 6L)

tailr(x, m = 6L, n = 6L)
}
\arguments{
\item{x}{A matrix or data frame.}

\item{m}{The number of rows to view. Default: 6.}

\item{n}{The number of columns to view. Default: 6.}
}
\value{
A matrix of size \code{m} x \code{n}.
}
\description{
Functions to view the top/bottom left/right corners of large matrices or data frames.
}
\examples{
mat <- matrix(1:900, 30, 30)
headl(mat)
tailr(mat)
headl(mat, 10, 5)
headl(as.data.frame(mat))
}
