% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_map.R
\name{heat_map}
\alias{heat_map}
\title{No-frills heatmap}
\usage{
heat_map(
  mat,
  ord = NULL,
  ord_x = NULL,
  ord_y = NULL,
  draw_dendro = FALSE,
  return_data = FALSE,
  show_legend = TRUE,
  plot_title = NULL,
  axis_text_x = colnames(mat),
  axis_text_y = rownames(mat),
  axis_title_x = waiver(),
  axis_title_y = waiver(),
  colours = c("red", "yellow"),
  limits = NULL,
  oob = scales::squish,
  legend_breaks = waiver(),
  legend_labels = waiver(),
  legend_title = waiver(),
  axis_ticks = element_blank(),
  axis_ticks_length = 0,
  axis_text_size = 7,
  plot_margin = c(5.5, 5.5, 5.5, 5.5),
  rel_widths = c(heatmap = 1, dend = 0.2, legend = 0.2),
  rel_heights = c(heatmap = 1, dend = 0.2),
  ...
)
}
\arguments{
\item{mat}{A numeric matrix.}

\item{ord}{An optional ordering object with which to reorder both the rows and the columns of \code{mat}. Can be an \code{hclust} object or an integer permutation vector.}

\item{ord_x}{An optional ordering object with which to reorder the columns of \code{mat}. Can be an \code{hclust} object or an integer permutation vector. Ignored if \code{ord} is supplied.}

\item{ord_y}{An optional ordering object with which to reorder the rows of \code{mat}. Can be an \code{hclust} object or an integer permutation vector. Ignored if \code{ord} is supplied.}

\item{draw_dendro}{Logical or character indicating whether to draw dendrograms or which dendrograms to draw. Possible values are \code{TRUE}, \code{FALSE}, 'x', 'y', 'xy'. If \code{TRUE} or 'xy', all possible dendrograms will be drawn. Default: \code{FALSE}.}

\item{return_data}{Logical indicating whether to return a data frame (in long format) containing the data used to construct the heatmap.}

\item{show_legend}{Logical indicating whether to show the heatmap legend.}

\item{plot_title}{The title of the plot. Set to \code{NULL} if you don't want a title. Default: \code{NULL}.}

\item{axis_text_x}{Character vector of x axis labels. These will be re-ordered if \code{ord} or \code{ord_x} are supplied. Set to \code{NULL} if you want no x axis labels. Default: \code{colnames(mat)}.}

\item{axis_text_y}{Character vector of y axis labels. These will be re-ordered if \code{ord} or \code{ord_y} are supplied. Set to \code{NULL} if you want no y axis labels. Default: \code{rownames(mat)}.}

\item{axis_title_x}{The x axis title. Set to \code{NULL} if you want no x axis title.}

\item{axis_title_y}{The y axis title. Set to \code{NULL} if you want no y axis title.}

\item{colours}{The colours to be used in the heatmap. These will be converted to a colour scale by ggplot2's \code{scale_fill_gradientn}. Default: c('red', 'yellow').}

\item{limits}{Two-element numeric vector specifying the limits to use for the colour scale.}

\item{oob}{A function to handle values lying outside the range specified by \code{limits}. Default: \code{scales::squish}.}

\item{legend_breaks}{A numeric vector of legend breaks.}

\item{legend_labels}{A vector of legend labels.}

\item{legend_title}{The legend title. Set to \code{NULL} if you want no legend title.}

\item{axis_ticks}{As \code{axis.ticks} in \code{theme}.}

\item{axis_ticks_length}{As \code{axis.ticks.length} in \code{theme}.}

\item{axis_text_size}{The size of the axis text. Default: 7.}

\item{plot_margin}{Numeric vector of plot margins in pts, in the order top, right, bottom, left. Default: c(5.5, 5.5, 5.5, 5.5).}

\item{rel_widths}{Named vector of relative widths to be passed to cowplot's \code{plot_grid}. Names should include 'heatmap' and, if applicable, 'dend' and 'legend'.}

\item{rel_heights}{Named vector of relative heights to be passed to cowplot's \code{plot_grid}. Names should include 'heatmap' and, if applicable, 'dend'.}

\item{...}{Additional arguments to be passed to ggplot2's \code{theme} function in the construction of the heatmap.}
}
\value{
If \code{return_data} is \code{FALSE}, a \code{ggplot} object. If \code{return_data} is \code{TRUE}, a list with two elements: \code{plot}, a \code{ggplot} object; and \code{data}, a data frame.
}
\description{
Makes a heatmap with optional row/column ordering and aligned dendrograms.
}
