% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/short_functions.R
\name{slapply}
\alias{slapply}
\title{lapply that uses names}
\usage{
slapply(X, FUN, ...)
}
\arguments{
\item{FUN}{The function to be applied to each element of \code{x}.}

\item{...}{Optional arguments to \code{FUN}.}

\item{x}{A vector (atomic or list).}
}
\value{
A list with names equal to the names of \code{x}.
}
\description{
A wrapper for \code{sapply} with arguments \code{simplify = FALSE} and \code{USE.NAMES = TRUE}.
}
