% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_matrix_functions.R
\name{to_frac}
\alias{to_frac}
\title{Convert to fractions of row or column totals}
\usage{
to_frac(mat, MARGIN = 2)
}
\arguments{
\item{mat}{A matrix. May be sparse (i.e. of class \code{dgTMatrix}, \code{dgCMatrix} or \code{dgRMatrix}).}

\item{MARGIN}{The subscript to apply over. Possible values are \code{1} and \code{2}. If \code{1}, each element will be divided by its corresponding row total; if \code{2}, its column total.}
}
\value{
A matrix of the same size and class as \code{mat}.
}
\description{
Divides the elements of a matrix by their corresponding row or column totals. This is useful, for example, when converting a gene expression count matrix to transcripts per million (TPM).
}
\examples{
mat <- matrix(round(runif(100, max = 100)), 10, 10)
to_frac(mat)
to_frac(mat, margin = 1)
mat[sample(1:100, 80)] <- 0
mat <- as(mat, 'dgCMatrix')
to_frac(mat)
}
